fix the ethers add wagmi
