+++ integrations/gitcoin-passport/components/score-stamp.tsx
@@ -1,15 +1,15 @@
 import { useState } from 'react'
 
-import { env } from '@/env.mjs'
 import { useToast } from '@/lib/hooks/use-toast'
 
-import { checkConnection } from '../gitcoin-passport-client'
-import { getScore } from '../hooks/use-get-score'
-import { ScoreGateProps } from '../utils/types'
+import { CheckConnection } from '../gitcoin-passport-client'
+import { getStamps } from '../hooks/use-get-stamp'
+import { ScoreStampProps } from '../utils/types'
 
-export function ScoreGate({ setScore }: ScoreGateProps) {
+export function ScoreStamp({ setStamps }: ScoreStampProps) {
   const [isLoading, setIsLoading] = useState<boolean>(false)
   const { toast, dismiss } = useToast()
+  const address = CheckConnection()
 
   const handleToast = ({ title, description }: { title: string; description: string }) => {
     toast({
@@ -24,20 +24,17 @@ export function ScoreGate({ setScore }: ScoreGateProps) {
 
   const handleGetScore = async () => {
     setIsLoading(true)
-    setScore(null)
     try {
-      const address = await checkConnection()
       if (address) {
-        console.log('api key', env.NEXT_PUBLIC_GITCOIN_PASSPORT_API_KEY)
-        const score = await getScore(address)
-        if (score && typeof score === 'number') {
-          setScore(score)
+        const stamps = await getStamps(address)
+        if (stamps) {
+          setStamps(stamps)
         }
       }
     } catch (e) {
       handleToast({
         title: 'An Error Occurred',
-        description: 'An error occurred while getting score.',
+        description: 'An error occurred while getiing stamps',
       })
     } finally {
       setIsLoading(false)
@@ -45,9 +42,9 @@ export function ScoreGate({ setScore }: ScoreGateProps) {
   }
 
   return (
-    <div className="m-4">
+    <div className="my-4 ml-4">
       <button className="btn btn-emerald" disabled={isLoading} onClick={handleGetScore}>
-        {isLoading ? 'Getting Score...' : 'Get Score'}
+        {isLoading ? 'Getting Stamps...' : 'Get Stamps'}
       </button>
     </div>
   )
